+++ provider/cmd/pulumi-tfgen-minio/main.go
@@ -15,12 +15,12 @@
 package main
 
 import (
+	minio "github.com/pulumi/pulumi-minio/provider"
+	"github.com/pulumi/pulumi-minio/provider/pkg/version"
 	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfgen"
-	xyz "github.com/pulumi/pulumi-xyz/provider"
-	"github.com/pulumi/pulumi-xyz/provider/pkg/version"
 )
 
 func main() {
 	// Modify the path to point to the new provider
-	tfgen.Main("xyz", version.Version, xyz.Provider())
+	tfgen.Main("minio", version.Version, minio.Provider())
 }