+++ sdk/go/minio/doc.go
@@ -1,0 +1,3 @@
+// A Pulumi package for creating and managing minio cloud resources.
+//
+package minio