+++ CHANGELOG.md
@@ -2,5 +2,6 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
+* Initial scaffolding of the provider against v1.2.0 of the Minio Terraform Provider.
 
 ---